+++ httpproxy/__init__.py
@@ -1,4 +1,4 @@
-""""
+"""
 Django HTTP Proxy - A simple HTTP proxy for the Django framework.
 """
 from ._version import get_versions